+++ application/models/Movements.php
@@ -1,10 +1,14 @@
 <?php
 
-class Movements extends MY_Model
+class Movements extends MY_Model2
 {
     // constructor
     function __construct()
     {
-        parent::__construct('movements');
+        parent::__construct('movements', 'Code', 'Datetime');
+    }
+
+    function displayMovements($code){
+        return $this->group($code);
     }
 }