+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.0.4 - 2016-02-11
+### Added:
+- Added stock/display
+- Added dropdown list in stock index
+- Altered Movements and Transactions models
+- Altered autolaod
+    - autoloaded movements and transactions models
+    - url helper
+
+
+
 ## 1.0.3 - 2016-02-10
 ### Added:
 - Added stockcontroller